requirements.txt
pytest==8.1.1
selenium==4.19.0
webdriver-manager==4.0.1
python-dotenv==1.0.1

#person that will pull our project will run
# pip install -r requirements.txt
#
